moving forward with numpy solution for analysis_sorter
