+++ test/test_utils.py
@@ -1,1 +1,4 @@
 import pytest
+
+def test_save_object_np_correctly_applies_rows_and_columns():
+    pass
